refactor(music): combine `join` and `play` command so that `play` auto-joins
