+++ src/commands/mod.rs
@@ -10,7 +10,7 @@ mod ping;
 pub use ping::ping;
 mod play;
 pub use play::play;
-mod join;
-pub use join::join;
 mod leave;
 pub use leave::leave;
+mod skip;
+pub use skip::skip;
